Finished Input Player and Pawns Guide
I've learned about Pawns is a type of Actor that's controlled by a human player or AI. I also learned the difference between the Actions Mapping and the Axis Mappings, where the Actions Mapping are for the "yes or no" inputs, while the Axis Mapping is for the "how much" input per frame. I bind those inputs at the Project Configuration. After that, I made a function to set response for the described mappings. While preparing that, I learned Clamp from FMath class, where you can keep the range between -1 to 1 while the player pressing couple buttons together that has the same function.